refactor: rename auth directory to user
