+++ src/adapters/api/create-user-json.presenter.ts
@@ -1,9 +1,9 @@
 import type { FastifyReply } from 'fastify';
 
-import type { CreateUserResponse } from '../../core/auth/dto/create-user.response';
-import { UserEmailInvalidError } from '../../core/auth/error/user-email-invalid.error';
-import { UserPasswordInvalidError } from '../../core/auth/error/user-password-invalid.error';
 import type { Presenter } from '../../core/common/presenter.interface';
+import type { CreateUserResponse } from '../../core/user/dto/create-user.response';
+import { UserEmailInvalidError } from '../../core/user/error/user-email-invalid.error';
+import { UserPasswordInvalidError } from '../../core/user/error/user-password-invalid.error';
 import { isInstanceOf } from '../../utils/instanceof.util';
 
 export class CreateUserJSONPresenter implements Presenter<CreateUserResponse> {
